WAT-480 refactor windowFeature to @testring/types
