+++ packages/types/src/web-application/index.ts
@@ -28,3 +28,21 @@ export interface IWebApplicationResponseMessage {
 export type IWebApplicationClient = {
     [K in keyof IBrowserProxyPlugin]: (...args: Array<any>) => Promise<any>
 };
+
+export type WindowFeatureBoolean = 'yes' | 'no';
+
+export type WindowFeaturesConfig = string | IWindowFeatures;
+
+export interface IWindowFeatures {
+    top?: number;
+    left?: number;
+    width?: number;
+    height?: number;
+    status?: WindowFeatureBoolean;
+    toolbar?: WindowFeatureBoolean;
+    menubar?: WindowFeatureBoolean;
+    location?: WindowFeatureBoolean;
+    resizable?: WindowFeatureBoolean;
+    scrollbars?: WindowFeatureBoolean;
+}
+
